Cast the minutes and seconds value to LONG before using them in the .Format() call.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@9832 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/SVNProgressDlg.cpp
@@ -1286,7 +1286,7 @@ UINT CSVNProgressDlg::ProgressThread()
 	{
 		CString sFinalInfo;
 		CTimeSpan time = CTime::GetCurrentTime() - startTime;
-		temp.Format(IDS_PROGRS_TIME, time.GetTotalMinutes(), time.GetSeconds());
+		temp.Format(IDS_PROGRS_TIME, (LONG)time.GetTotalMinutes(), (LONG)time.GetSeconds());
 		sFinalInfo.Format(IDS_PROGRS_FINALINFO, m_sTotalBytesTransferred, temp);
 		GetDlgItem(IDC_PROGRESSLABEL)->SetWindowText(sFinalInfo);
 	}
